requirements.txt
boto3-1.24.27
requests-2.28.1
requests_oauthlib-1.3.1